refactor: 광고 예측 기능 predict_ad() 함수로 분리 및 구조 정리
- services/predict_ad.py 에서 CLI 실행 로직을 분리하여 cli_predict_ad.py 로 이동
- predict_ad() 함수 정의 및 Whisper 서비스에서 직접 호출할 수 있도록 모듈화
- whisper_service.py 에서 기존 ad_predictor import → predict_ad 로 수정
- 전체 광고 예측 코드 구조를 CLI와 서비스 호출 용도로 명확히 분리
